Move theme-dependent view spec into main project
